+++ CHANGELOG.md
@@ -36,3 +36,4 @@ Config-specific:
 - [#42] Implement config repositories
 - [#48] Save config diff in cluster
 - [#39] Encrypt sensitive data
+- [#36] Set registry configuration for dogu and global config.